requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Twilio
twilio==8.10.0

# Audio processing
numpy==1.24.3
scipy==1.11.4
soundfile==0.12.1
librosa==0.10.1
webrtcvad==2.0.10
pydub==0.25.1
pyaudio==0.2.14

# Speech recognition (Whisper)
faster-whisper==0.10.0
openai-whisper==20231117

# Text-to-Speech
TTS==0.21.0  # Coqui TTS for multiple engines
piper-tts==1.2.0

# LLM
ollama==0.1.6
langchain==0.1.0
langchain-community==0.0.10

# Async and concurrency
aiohttp==3.9.1
asyncio==3.4.3
aioconsole==0.6.2
aiofiles==23.2.1

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0
python-json-logger==2.0.7

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Performance
numba==0.58.1
onnxruntime-gpu==1.16.3  # For GPU acceleration

# Utilities
pyyaml==6.0.1
redis==5.0.1  # For session management
python-multipart==0.0.6
httpx==0.25.2

# Echo cancellation
pyAEC==0.3.0  # Acoustic echo cancellation
noisereduce==3.0.0

# Additional audio tools
pedalboard==0.8.1  # Audio effects processing
